daemon/main: Remove waitFDPair and related blocking code on it
In a Unix-style daemon the parent process should *immediately* exit after
it realizes it is set to daemonize. The current behavior was causing the
parent process to terminate only after the child was finised loading
all of its counters, leading to bizarre behavior when being started using
Ubuntu's Upstart.

As is customary with Unix-style daemons, this patch also introduces
the child process correctly creating its own process group using
the setsid(2) call.