Add Travis-CI support
Most of the changes were made to allow this code to be
built on travis-ci, a free continous integration service.
Notably, the code had to make sure to close stdin when
forking a process. In one of the tests comparing json
documents, the test was changed to use a sorted csv of
json properties. A direct comparison using just diff produces
wrong results, as json properties are unordered.